refactor: 코드 구조 개선을 위해 렌더링 관련 함수들을 별도의 모듈로 분리
